+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQ9J5A83R0YFX465DR7TX400"
+        "01KQ9RQW9P441D62DWGWNQG7CH"
     )
     r"""The unique identifier of the data source"""