+++ packages/orq-rc/src/orq_ai_sdk/models/listidentitiesop.py
@@ -208,7 +208,7 @@ class ListIdentitiesData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-28T08:09:01.996Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-28T10:04:02.081Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")