+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-28T08:09:01.996Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-28T10:04:02.081Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")