+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQ9J5A85TE9MAB93RJQ0C34F"
+        "01KQ9RQW9RRMCCCPE950A5XTE5"
     )
     r"""The unique identifier of the data source"""