+++ packages/orq-rc/RELEASES.md
@@ -7379,3 +7379,13 @@ Based on:
 - [python v4.9.0-rc.5] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.5 - packages/orq-rc
+
+## 2026-04-28 10:08:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.6] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.6] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.6 - packages/orq-rc